+++ crates/deslop-lsp/tests/execute_command.rs
@@ -18,7 +18,7 @@ use tower_lsp::{
 };
 
 use crate::common::{
-    call, copy_fixture, handshake, read_frame, request, spawn_lsp, take_io, write_frame,
+    call, copy_fixture, handshake, read_frame, request, spawn_lsp_on_fixture, write_frame,
 };
 use deslop_lsp::LspBackend;
 
@@ -26,9 +26,8 @@ const EXECUTE_COMMAND: &str = "workspace/executeCommand";
 
 #[test]
 fn execute_command_provider_advertises_and_opens_virtual_documents() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (_workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let init = handshake(&mut stdin, &mut stdout)?;
     let commands = advertised_commands(&init)?;
 
@@ -83,9 +82,8 @@ fn execute_command_provider_advertises_and_opens_virtual_documents() -> Result<(
 
 #[test]
 fn execute_command_dispatches_refresh_models_and_incremental_toggle() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (_workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
 
     let initial_config = call(&mut stdin, &mut stdout, "deslop/sessionConfig", &json!({}))?;
