+++ crates/deslop-lsp/tests/virtual_document.rs
@@ -12,16 +12,15 @@ use std::{path::Path, thread, time::Duration};
 use anyhow::{anyhow, Result};
 use serde_json::{json, Value};
 
-use crate::common::{call, copy_fixture, handshake, notification, spawn_lsp, take_io, write_frame};
+use crate::common::{call, handshake, notification, spawn_lsp_on_fixture, write_frame};
 
 const VIRTUAL_DOCUMENT: &str = "deslop/virtualDocument";
 const REPORT_GET: &str = "deslop/reportGet";
 
 #[test]
 fn virtual_document_schema_returns_non_empty_markdown() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (_workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
 
     let response = call(
@@ -44,9 +43,8 @@ fn virtual_document_schema_returns_non_empty_markdown() -> Result<()> {
 
 #[test]
 fn virtual_document_report_returns_canonical_text() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
     open_fixture_files(&mut stdin, workspace.path())?;
 
@@ -77,9 +75,8 @@ fn virtual_document_report_returns_canonical_text() -> Result<()> {
 
 #[test]
 fn virtual_document_cluster_returns_cluster_markdown() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
     open_fixture_files(&mut stdin, workspace.path())?;
     let cluster_id = wait_for_first_cluster(&mut stdin, &mut stdout)?;
@@ -106,9 +103,8 @@ fn virtual_document_cluster_returns_cluster_markdown() -> Result<()> {
 
 #[test]
 fn virtual_document_rejects_malformed_uri_with_invalid_params() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (_workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
 
     let response = call(
@@ -132,9 +128,8 @@ fn virtual_document_rejects_malformed_uri_with_invalid_params() -> Result<()> {
 
 #[test]
 fn virtual_document_rejects_unknown_cluster_id() -> Result<()> {
-    let workspace = copy_fixture("csharp-small")?;
-    let mut child = spawn_lsp(workspace.path())?;
-    let (mut stdin, mut stdout, _stderr) = take_io(&mut child)?;
+    let (_workspace, mut child, mut stdin, mut stdout, _stderr) =
+        spawn_lsp_on_fixture("csharp-small")?;
     let _init = handshake(&mut stdin, &mut stdout)?;
 
     let response = call(
