+++ crates/deslop-mcp/tests/issue_137_ipc_freshness.rs
@@ -14,8 +14,9 @@ use serde_json::{json, Value};
 
 mod common;
 use common::{
-    copied_fixture, initialized_mcp, spawn_lsp_and_initialize, structured_content, wait_for_path,
-    ChildKillOnDrop, McpHandle, SOCKET_TIMEOUT,
+    cluster_ids, copied_fixture, initialized_mcp, lsp_workspace_with_socket,
+    spawn_lsp_and_initialize, structured_content, wait_for_path, ChildKillOnDrop, McpHandle,
+    SOCKET_TIMEOUT,
 };
 
 /// [MCP-IPC-CLIENT] T1 — read freshness without on-disk staleness.
@@ -28,11 +29,7 @@ use common::{
 /// updates no longer rewrite it ([LIVE-SEED-CACHE]).
 #[test]
 fn t1_report_get_reflects_lsp_state_immediately_after_rescan() -> Result<()> {
-    let workspace = copied_fixture()?;
-    let lsp = spawn_lsp_and_initialize(workspace.path())?;
-    let _lsp_guard = ChildKillOnDrop(lsp);
-    let socket = workspace.path().join(".deslop-cache/deslop.sock");
-    wait_for_path(&socket, SOCKET_TIMEOUT).context("wait for ipc socket")?;
+    let (workspace, _lsp_guard, _socket) = lsp_workspace_with_socket()?;
     let mut mcp = initialized_mcp(workspace.path())?;
 
     let baseline = call_report_get(&mut mcp)?;
@@ -186,18 +183,4 @@ fn call_report_get(mcp: &mut McpHandle) -> Result<Value> {
         }),
     )?;
     structured_content(&response, "report-get")
-}
-
-/// Returns the cluster ids on a `report-get` page in stable order.
-fn cluster_ids(page: &Value) -> Vec<String> {
-    page.get("clusters")
-        .and_then(Value::as_array)
-        .map(|clusters| {
-            clusters
-                .iter()
-                .filter_map(|cluster| cluster.get("id").and_then(Value::as_str))
-                .map(str::to_owned)
-                .collect()
-        })
-        .unwrap_or_default()
 }
