+++ crates/deslop-mcp/tests/state_file_backend.rs
@@ -28,8 +28,8 @@ use serde_json::{json, Value};
 
 mod common;
 use common::{
-    copied_fixture, initialized_mcp, spawn_lsp_and_initialize, structured_content, wait_for_path,
-    ChildKillOnDrop, SOCKET_TIMEOUT,
+    cluster_ids, copied_fixture, initialized_mcp, lsp_workspace_with_socket,
+    spawn_lsp_and_initialize, structured_content, wait_for_path, ChildKillOnDrop, SOCKET_TIMEOUT,
 };
 
 /// [MCP-IPC-CLIENT] Repurposes `issue_90_report_get_reloads_state_file_between_plain_calls`.
@@ -45,11 +45,7 @@ use common::{
 /// state cannot survive a forced LSP refresh.
 #[test]
 fn issue_90_report_get_reflects_lsp_state_between_plain_calls() -> Result<()> {
-    let workspace = copied_fixture()?;
-    let lsp = spawn_lsp_and_initialize(workspace.path())?;
-    let _lsp_guard = ChildKillOnDrop(lsp);
-    let socket = workspace.path().join(".deslop-cache/deslop.sock");
-    wait_for_path(&socket, SOCKET_TIMEOUT).context("wait for ipc socket")?;
+    let (workspace, _lsp_guard, _socket) = lsp_workspace_with_socket()?;
 
     let mut mcp = initialized_mcp(workspace.path())?;
 
@@ -202,18 +198,4 @@ fn call_report_get(mcp: &mut common::McpHandle, offset: u64, limit: u64) -> Resu
         }),
     )?;
     structured_content(&response, "report-get")
-}
-
-/// Returns the cluster ids on a `report-get` page in stable order.
-fn cluster_ids(page: &Value) -> Vec<String> {
-    page.get("clusters")
-        .and_then(Value::as_array)
-        .map(|clusters| {
-            clusters
-                .iter()
-                .filter_map(|cluster| cluster.get("id").and_then(Value::as_str))
-                .map(str::to_owned)
-                .collect()
-        })
-        .unwrap_or_default()
 }
