+++ crates/deslop/tests/csharp_type1_type2_distinct_buckets.rs
@@ -7,20 +7,13 @@
 //! (two methods that are byte-identical) MUST have at least one cluster
 //! claiming "every copy is the same".
 
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use std::{fs, path::Path};
 
 use anyhow::Result;
 use assert_cmd::Command;
 
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join(name)
-}
+mod common;
+use crate::common::*;
 
 fn run_report(tmp: &Path, scan_root: &Path) -> Result<serde_json::Value> {
     let mut cmd = Command::cargo_bin("deslop")?;
