+++ crates/deslop/tests/dart_issue_197_single_file_structural_only.rs
@@ -17,21 +17,14 @@
 //! meilisearch-dart settings region, so the cluster ids `be951a686525` and
 //! `7f363063109f` from the issue reproduce verbatim.
 
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use std::{fs, path::Path};
 
 use anyhow::Result;
 use assert_cmd::Command;
 use serde_json::Value;
 
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join(name)
-}
+mod common;
+use crate::common::*;
 
 fn run_report(scan_root: &Path) -> Result<Value> {
     let tmp = tempfile::tempdir()?;
