+++ crates/deslop/tests/issue_168_deep_nesting_no_crash.rs
@@ -16,13 +16,8 @@ use anyhow::Result;
 use assert_cmd::Command;
 use serde_json::Value;
 
-fn clusters(report: &Value) -> &[Value] {
-    report
-        .get("clusters")
-        .and_then(Value::as_array)
-        .map(Vec::as_slice)
-        .unwrap_or_default()
-}
+mod common;
+use crate::common::*;
 
 fn report_path(tmp: &Path) -> PathBuf {
     let mut path = tmp.join("report");
