+++ crates/deslop/tests/python_dict_false_positive.rs
@@ -16,17 +16,13 @@
 //! sit in the same file with non-overlapping but adjacent byte ranges
 //! inside the same parent node must be suppressed.
 
-use std::{fs, path::Path, path::PathBuf};
+use std::fs;
 
 use anyhow::Result;
 use assert_cmd::Command;
 
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join(name)
-}
+mod common;
+use crate::common::*;
 
 fn run_cli_on_fixture(fixture_name: &str) -> Result<serde_json::Value> {
     let tmp = tempfile::tempdir()?;
