+++ crates/deslop/tests/python_issue_72_monkeypatch.rs
@@ -1,20 +1,13 @@
 //! E2E regression for GH #72: `monkeypatch.setenv` setup patterns across
 //! config tests are scaffolding, not duplicate logic.
 
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use std::fs;
 
 use anyhow::Result;
 use assert_cmd::Command;
 
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join(name)
-}
+mod common;
+use crate::common::*;
 
 fn run_report(fixture_name: &str) -> Result<serde_json::Value> {
     let tmp = tempfile::tempdir()?;
