+++ crates/deslop/tests/rust_issue_176_match_dispatch.rs
@@ -7,21 +7,14 @@
 //! must never surface in the report.
 //! Spec: [CLONE-NOISE-RUST-MATCH-DISPATCH].
 
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use std::fs;
 
 use anyhow::Result;
 use assert_cmd::Command;
 use serde_json::Value;
 
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join(name)
-}
+mod common;
+use crate::common::*;
 
 fn run_report(fixture_name: &str) -> Result<Value> {
     let tmp = tempfile::tempdir()?;
@@ -88,11 +81,4 @@ fn rust_verbatim_copied_match_arms_still_cluster() -> Result<()> {
          genuine duplication and must still surface as a cluster: {report:#}"
     );
     Ok(())
-}
-
-fn clusters(report: &Value) -> &[Value] {
-    report
-        .get("clusters")
-        .and_then(Value::as_array)
-        .map_or(&[][..], Vec::as_slice)
 }
